+++ CuriousX/Parser/README.md
@@ -1,0 +1,62 @@
+## Syntax Analysis
+
+The syntax analyzer takes tokens from the lexer as input, it checks if the sequence of tokens generated by the lexer is a valid program according to the language's grammar and outputs a parse tree 
+
+Grammar rules in natural language :
+
+-   Expressions can consist of operands (such as numbers or variables) and operators (such as +, -, *, /).
+-   Parentheses can be used to group sub-expressions and indicate precedence.
+-   Operators have a specific order of precedence (such as multiplication and division before addition and subtraction).
+-   Operators with the same precedence have a specific associativity (such as left-to-right or right-to-left).
+
+In a formal notation
+
+```c++
+<expr> ::= <term> {<addop> <term>}
+
+<term> ::= <factor> {<mulop> <factor>}
+
+<factor> ::= '(' <expr> ')' | <number> | <identifier>
+
+<addop> ::= '+' | '-'
+
+<mulop> ::= '*' | '/'
+```
+
+In summary 
+
+-   Expression is made up of sum of terms
+-   Term is a product of factors
+-   Factor is a number or parenthesized sub expression
+
+The `Node.hpp` file contains utility function for the tree nodes
+
+```c++
+// Node structure
+struct Node
+{
+  std::shared_ptr<Node> left;
+  std::shared_ptr<Node> right;
+  LexerToken type;
+};
+```
+For example, consider the following expression in our CuriousX:
+
+```c
+x = 5 + 2 * 3
+```
+
+The parse tree for this expression would have the following structure:
+```sh
+                         =
+                        / \
+                       x   +
+                          / \
+                         5   *
+                            / \
+                           2   3
+
+```
+
+__To learn more:__
+## [CuriousX Syntax Analyzer Blog](https://jenniferchukwu.com/posts/syntaxanalysis)
